メイドちゃんを今後拡張しやすいようにリファクタリング（http部分）
